+++ src/ChainState.js
@@ -2,8 +2,10 @@ import React, { useEffect, useState } from "react";
 
 import { Grid, Form, Dropdown, Button, Input } from "semantic-ui-react";
 
+import { useSubstrate } from "./substrate";
+
 export default function ChainState(props) {
-  const { api } = props;
+  const { api } = useSubstrate();
 
   const [modulesList, setModulesList] = useState([]);
   const [output, setOutput] = useState("");
