+++ src/Events.js
@@ -1,9 +1,10 @@
 import React, { useEffect, useState } from "react";
-
 import { Feed, Grid } from "semantic-ui-react";
 
+import { useSubstrate } from "./substrate";
+
 export default function Events(props) {
-  const { api } = props;
+  const { api } = useSubstrate();
 
   const [eventFeed, setEventFeed] = useState([]);
