+++ src/examples/ProofOfExistence.js
