+++ src/examples/TemplateModule.js
