+++ src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 
 import App from "./App";
-import { SubstrateContextProvider } from "./substrate/SubstrateContext";
+import { SubstrateContextProvider } from "./substrate";
 
 ReactDOM.render(
   // More information on custom types: http://bit.ly/2kHYvLc
