+++ src/substrate/components/index.js
@@ -1,0 +1,3 @@
+import DeveloperConsole from "./DeveloperConsole";
+
+export { DeveloperConsole };
