+++ src/substrate/index.js
@@ -1,0 +1,6 @@
+import useSubstrate from "./useSubstrate";
+import {
+  SubstrateContext, SubstrateContextProvider
+} from "./SubstrateContext";
+
+export { useSubstrate, SubstrateContext, SubstrateContextProvider };
